{"types":"C","runs":3}

+++ Problem15.java
@@ -4,6 +4,6 @@ public class Problem15 extends ConsoleProgram {
 	public void run() {
 		int x = readInt("Enter the first number: ");
 		int y = readInt("Enter the first number: ");
-		println("the arithmetic mean of these numbers is: " + ( x + y ) / 2);
+		println("the arithmetic mean of these numbers is: " + (double) ((x + y) / 2));
 	}
 }
